add cooccurence stuff

+++ libs/visualisation/visualisation_KNN.m
@@ -2,9 +2,8 @@
   % formatting output data 
   %generate the clusters struct
   clusters = struct();
-  
   patches_to_crops = [];
-  
+
   for (i = 1 : numel(best_clusters_idx))
   
     clusters(i).id = best_clusters_idx(i);
@@ -14,6 +13,9 @@
     
     centroid.position = centroid_pos(:);
     centroid.img_path = [centroid_pos(1)  i -1];
+    % patch level
+    
+    centroid.level = patch_size(centroid.position(2:5)') /ds.params.patchCanonicalSize(1);
     clusters(i).centroid = centroid;
     
     tmp = [centroid_pos i -1];
@@ -30,28 +32,38 @@
      
      end
     clusters(i).nn = nn;
-
+ 
     % purity 
     clusters(i).purity = purity(best_clusters_idx(i));
   end
 
- experiment_dir = sprintf('results/%s',ds.params.experiment_name);
 
+create = true;
+ root_dir = sprintf('results/%s/nn',ds.params.experiment_name);
+ 
+if (create)
+  % clean dir 
+   if (exist(root_dir))
+    rmdir(root_dir,'s');
+   end 
+   mkdir(root_dir);
+end
 
-% create dir to save results from this experiment
-root_dir = sprintf('results/%s/nn',ds.params.experiment_name);
- if (exist(root_dir))
-  rmdir(root_dir,'s')
- end 
-mkdir(root_dir);
-
-% extract images
-img_dir = [root_dir '/images'];
-mkdir(img_dir);
-
-save_img_patches(patches_to_crops, imgs, img_dir);
-%save_img_patches([clusters.nn], imgs, img_dir);
 
 % save clusters.json
+
 json_file = [root_dir '/clusters_knn.json'];
 savejson('',clusters,json_file);
+
+if (create)
+  % extract images
+  img_dir = [root_dir '/images'];
+  mkdir(img_dir);
+  save_img_patches(patches_to_crops, imgs, img_dir);
+  
+  %save_img_patches([clusters.nn], imgs, img_dir);
+end
+
+
+
+
